pyproject.toml
[project]
name = "phonagnosia"
version = "0.1.0"
description = "Voice-first multilingual assistant demo with Siemens LLM/STT, Streamlit, MCP, and ElevenLabs TTS"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "elevenlabs>=2.39.0",
    "fastapi>=0.135.1",
    "fastmcp>=3.1.1",
    "json-repair>=0.58.5",
    "mcp>=1.26.0",
    "openai>=2.28.0",
    "pydantic>=2.11.7",
    "pyyaml>=6.0.3",
    "streamlit>=1.55.0",
    "uvicorn>=0.41.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]

[tool.pytest.ini_options]
pythonpath = ["."]
